+++ NEWS.md
@@ -15,10 +15,12 @@ Version 0.15a0 - dev
 - NEW: `Circle.flattening()` approximation determined by a max. sagitta value
 - NEW: `Arc.flattening()` approximation determined by a max. sagitta value
 - NEW: `ConstructionArc.flattening()` approximation determined by a max. sagitta value
+- NEW: `ezdxf.math.distance_point_line_3d()`
+- NEW: `ConstructionEllipse.flattening()` adaptive recursive flattening (approximation)
+- NEW: `Ellipse.flattening()` adaptive recursive flattening (approximation)
 - NEW: `matplotlib.qsave()`, `ltype` argument to switch between matplotlib dpi 
   based linetype rendering and AutoCAD like drawing units based linetype 
   rendering
-- NEW: `ezdxf.math.distance_point_line_3d()`
 - REMOVED: deprecated `DXFEntity.transform_to_wcs()` interface, 
   use `DXFEntity.transform(ucs.matrix)`
 - REMOVED: deprecated `Hatch.edit_boundary()` context manager,